+++ Views/GameData/Edit/RoomDetailDialog.axaml.cs
@@ -1,0 +1,11 @@
+using Avalonia.Controls;
+
+namespace FujinTerm.Views.GameData.Edit;
+
+public partial class RoomDetailDialog : Window
+{
+    public RoomDetailDialog()
+    {
+        InitializeComponent();
+    }
+}